refact: 이벤트 핸들러를 features로 분리
