+++ src/basic/main.basic.js
@@ -7,6 +7,7 @@ import { initializeApplication, initializeProductData } from './features/initial
 import { createDOMElements } from './widgets/dom-creator/index.js';
 import { setupPromotionTimers } from './features/promotion/index.js';
 import { calculateCartSubtotal, calculateFinalDiscount, updateCartUI } from './features/pricing/index.js';
+import { onUpdateSelectOptions, updateStockStatus, handleCalculateCartStuff, doRenderBonusPoints, onGetStockTotal, doUpdatePricesInCart } from './features/events/index.js';
 
 // 레거시 호환성을 위한 AppState 참조
 const AppState = appState;
@@ -74,10 +75,34 @@ function setupPromotionTimersLegacy() {
   setupPromotionTimers(AppState, legacyVars);
 }
 
+function onUpdateSelectOptionsLegacy() {
+  onUpdateSelectOptions(AppState, legacyVars);
+}
+
+function updateStockStatusLegacy() {
+  updateStockStatus(AppState, legacyVars);
+}
+
+function handleCalculateCartStuffLegacy() {
+  handleCalculateCartStuff(AppState, legacyVars);
+}
+
+function doRenderBonusPointsLegacy() {
+  doRenderBonusPoints(AppState, legacyVars);
+}
+
+function onGetStockTotalLegacy() {
+  return onGetStockTotal(AppState);
+}
+
+function doUpdatePricesInCartLegacy() {
+  doUpdatePricesInCart(AppState, legacyVars);
+}
+
 function initializeUI() {
   var initStock = calculateTotalStockLegacy();
-  onUpdateSelectOptions();
-  handleCalculateCartStuff();
+  onUpdateSelectOptionsLegacy();
+  handleCalculateCartStuffLegacy();
 }
 
 function main() {
@@ -88,52 +113,6 @@ function main() {
   initializeUI();
 };
 var sum
-function onUpdateSelectOptions() {
-  var totalStock;
-  var opt;
-  var discountText;
-  AppState.elements.productSelect.innerHTML = '';
-  sel.innerHTML = ''; // 레거시 동기화
-  
-  totalStock = calculateTotalStockLegacy();
-  for (var i = 0; i < AppState.products.length; i++) {
-    (function() {
-      var item = AppState.products[i];
-      opt = document.createElement("option")
-      opt.value = item.id;
-      discountText = '';
-      if (item.onSale) discountText += ' ⚡SALE';
-      if (item.suggestSale) discountText += ' 💝추천';
-      if (item.q === 0) {
-        opt.textContent = item.name + ' - ' + item.val + '원 (품절)' + discountText
-        opt.disabled = true
-        opt.className = 'text-gray-400';
-      } else {
-        if (item.onSale && item.suggestSale) {
-          opt.textContent = '⚡💝' + item.name + ' - ' + item.originalVal + '원 → ' + item.val + '원 (25% SUPER SALE!)';
-          opt.className = 'text-purple-600 font-bold';
-        } else if (item.onSale) {
-          opt.textContent = '⚡' + item.name + ' - ' + item.originalVal + '원 → ' + item.val + '원 (20% SALE!)';
-          opt.className = 'text-red-500 font-bold';
-        } else if (item.suggestSale) {
-          opt.textContent = '💝' + item.name + ' - ' + item.originalVal + '원 → ' + item.val + '원 (5% 추천할인!)';
-          opt.className = 'text-blue-500 font-bold';
-        } else {
-          opt.textContent = item.name + ' - ' + item.val + '원' + discountText;
-        }
-      }
-      AppState.elements.productSelect.appendChild(opt);
-      sel.appendChild(opt); // 레거시 동기화
-    })();
-  }
-  if (totalStock < AppState.CONSTANTS.STOCK_WARNING_THRESHOLD) {
-    AppState.elements.productSelect.style.borderColor = 'orange';
-    sel.style.borderColor = 'orange'; // 레거시 동기화
-  } else {
-    AppState.elements.productSelect.style.borderColor = '';
-    sel.style.borderColor = ''; // 레거시 동기화
-  }
-}
 function calculateCartSubtotalLegacy() {
   return calculateCartSubtotal(AppState, legacyVars);
 }
@@ -146,156 +125,12 @@ function updateCartUILegacy(subTotal, itemDiscounts, discountInfo) {
   updateCartUI(AppState, legacyVars, subTotal, itemDiscounts, discountInfo);
 }
 
-function updateStockStatus() {
-  var stockMsg = '';
-  
-  for (var stockIdx = 0; stockIdx < AppState.products.length; stockIdx++) {
-    var item = AppState.products[stockIdx];
-    if (item.q < AppState.CONSTANTS.LOW_STOCK_THRESHOLD) {
-      if (item.q > 0) {
-        stockMsg = stockMsg + item.name + ': 재고 부족 (' + item.q + '개 남음)\n';
-      } else {
-        stockMsg = stockMsg + item.name + ': 품절\n';
-      }
-    }
-  }
-  
-  AppState.elements.stockInfo.textContent = stockMsg;
-  
-  // 레거시 변수 동기화
-  stockInfo.textContent = stockMsg;
-}
 
-function handleCalculateCartStuff() {
-  // 1. 장바구니 소계 및 개별 할인 계산
-  var subtotalResult = calculateCartSubtotalLegacy();
-  
-  // 2. 전체 할인 계산 (대량구매, 화요일)
-  var discountResult = calculateFinalDiscountLegacy(subtotalResult.subTotal);
-  
-  // 3. UI 업데이트
-  updateCartUILegacy(subtotalResult.subTotal, subtotalResult.itemDiscounts, discountResult);
-  
-  // 4. 재고 상태 업데이트
-  updateStockStatus();
-  
-  // 5. 포인트 계산
-  doRenderBonusPoints();
-}
-var doRenderBonusPoints = function() {
-  var basePoints;
-  var finalPoints;
-  var pointsDetail;
-  var hasKeyboard;
-  var hasMouse;
-  var hasMonitorArm;
-  var nodes;
-  if (cartDisp.children.length === 0) {
-    document.getElementById('loyalty-points').style.display = 'none';
-    return;
-  }
-  basePoints = Math.floor(AppState.totalAmount / 1000)
-  finalPoints = 0;
-  pointsDetail = [];
-  if (basePoints > 0) {
-    finalPoints = basePoints;
-    pointsDetail.push('기본: ' + basePoints + 'p');
-  }
-  if (new Date().getDay() === 2) {
-    if (basePoints > 0) {
-      finalPoints = basePoints * 2;
-      pointsDetail.push('화요일 2배');
-    }
-  }
-  hasKeyboard = false;
-  hasMouse = false;
-  hasMonitorArm = false;
-  nodes = cartDisp.children;
-  for (const node of nodes) {
-    var product = findProductByIdLegacy(node.id);
-    if (!product) continue;
-    if (product.id === AppState.PRODUCT_IDS.KEYBOARD) {
-      hasKeyboard = true;
-    } else if (product.id === AppState.PRODUCT_IDS.MOUSE) {
-      hasMouse = true;
-    } else if (product.id === AppState.PRODUCT_IDS.MONITOR_ARM) {
-      hasMonitorArm = true;
-    }
-  }
-  if (hasKeyboard && hasMouse) {
-    finalPoints = finalPoints + 50;
-    pointsDetail.push('키보드+마우스 세트 +50p');
-  }
-  if (hasKeyboard && hasMouse && hasMonitorArm) {
-    finalPoints = finalPoints + 100;
-    pointsDetail.push('풀세트 구매 +100p');
-  }
-  if (AppState.itemCount >= 30) {
-    finalPoints = finalPoints + 100;
-    pointsDetail.push('대량구매(30개+) +100p');
-  } else {
-    if (AppState.itemCount >= 20) {
-      finalPoints = finalPoints + 50;
-      pointsDetail.push('대량구매(20개+) +50p');
-    } else {
-      if (AppState.itemCount >= 10) {
-        finalPoints = finalPoints + 20;
-        pointsDetail.push('대량구매(10개+) +20p');
-      }
-    }
-  }
-  bonusPts = finalPoints;
-  var ptsTag = document.getElementById('loyalty-points');
-  if (ptsTag) {
-    if (bonusPts > 0) {
-      ptsTag.innerHTML = '<div>적립 포인트: <span class="font-bold">' + bonusPts + 'p</span></div>' +
-        '<div class="text-2xs opacity-70 mt-1">' + pointsDetail.join(', ') + '</div>';
-      ptsTag.style.display = 'block';
-    } else {
-      ptsTag.textContent = '적립 포인트: 0p';
-      ptsTag.style.display = 'block'
-    }
-  }
-}
-function onGetStockTotal() {
-  return calculateTotalStockLegacy();
-}
-function doUpdatePricesInCart() {
-  var cartItems = cartDisp.children;
-  
-  // 장바구니의 각 아이템에 대해 가격과 이름 업데이트
-  for (var i = 0; i < cartItems.length; i++) {
-    var itemId = cartItems[i].id;
-    var product = findProductByIdLegacy(itemId);
-    
-    if (product) {
-      var priceDiv = cartItems[i].querySelector('.text-lg');
-      var nameDiv = cartItems[i].querySelector('h3');
-      
-      // 세일 상태에 따른 가격 및 이름 표시
-      if (product.onSale && product.suggestSale) {
-        // 번개세일 + 추천할인
-        priceDiv.innerHTML = '<span class="line-through text-gray-400">₩' + product.originalVal.toLocaleString() + '</span> <span class="text-purple-600">₩' + product.val.toLocaleString() + '</span>';
-        nameDiv.textContent = '⚡💝' + product.name;
-      } else if (product.onSale) {
-        // 번개세일만
-        priceDiv.innerHTML = '<span class="line-through text-gray-400">₩' + product.originalVal.toLocaleString() + '</span> <span class="text-red-500">₩' + product.val.toLocaleString() + '</span>';
-        nameDiv.textContent = '⚡' + product.name;
-      } else if (product.suggestSale) {
-        // 추천할인만
-        priceDiv.innerHTML = '<span class="line-through text-gray-400">₩' + product.originalVal.toLocaleString() + '</span> <span class="text-blue-500">₩' + product.val.toLocaleString() + '</span>';
-        nameDiv.textContent = '💝' + product.name;
-      } else {
-        // 일반 가격
-        priceDiv.textContent = '₩' + product.val.toLocaleString();
-        nameDiv.textContent = product.name;
-      }
-    }
-  }
-  
-  // 전체 계산 업데이트
-  handleCalculateCartStuff();
-}
+
+
+
+
+
 main();
 
 // 장바구니 추가 버튼 이벤트 핸들러
@@ -373,7 +208,7 @@ addBtn.addEventListener("click", function () {
       itemToAdd.q--;
     }
     
-    handleCalculateCartStuff();
+    handleCalculateCartStuffLegacy();
     lastSel = selItem;
   }
 });
@@ -413,7 +248,7 @@ cartDisp.addEventListener("click", function (event) {
     }
     
     // UI 업데이트
-    handleCalculateCartStuff();
-    onUpdateSelectOptions();
+    handleCalculateCartStuffLegacy();
+    onUpdateSelectOptionsLegacy();
   }
 });
